legacy_and_archives/arsiv/sosyal_medya_ajani/requirements.txt
httpx
pillow
moviepy
apscheduler
